chore(swap): move refresh quote logic to top-level Swap component
