+++ ui/app/pages/swap/swap-content/index.js
@@ -1,1 +1,1 @@
-export { default } from './swap-content.container'
+export { default } from './swap-content.component'
